requirements.txt
# Core dependencies
requests==2.31.0
selenium==4.15.2
python-dotenv==1.0.0
pyyaml==6.0.1

# Web driver management
webdriver-manager==4.0.1

# Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0

# Data handling
pandas==2.1.4

# Utilities
fake-useragent==1.4.0